Функции генерации структур разнесены по разным файлам
