+++ src/app/pages/VariantEditor/utils/generateGraphVE.ts
@@ -1,0 +1,14 @@
+export function generateGraphVE(vertexAmount: number, edgesAmount: number) {
+    const vertices = Array.from(Array(vertexAmount).keys())
+        .map(e => `"${e}"`)
+        .join(",");
+    const edges = Array.from(Array(edgesAmount).keys())
+        .map(e => `{ "name": "${e}", "sources": "sourceName" , "target": "targetName" }`)
+        .join(",\n\t");
+    return `{ "type": "graph", "value": {\n` +
+        `\t"vertices": [${vertices}], \n` +
+        `\t"edges": \n` +
+        `\t[${edges}] \n` +
+        `   } \n` +
+        `}`;
+}
